+++ src/interfaces/gridOptions/gridOptions.interface.ts
@@ -1,4 +1,3 @@
-import {PagingPosition} from '../../misc/enums';
 import {GridPluginTypes} from '../gridPluginTypes/gridPluginTypes.interface';
 
 /**
@@ -10,12 +9,6 @@ export interface GridOptions
      * Indication whether grid should be initialized automaticaly during 'NgOnInit' phase
      */
     autoInitialize: boolean;
-
-    //TODO: remove? or move
-    /**
-     * Position of paging rendered by grid
-     */
-    pagingPosition: PagingPosition;
 
     /**
      * Object defining overrides for default plugins, default plugins can be also specified using DI