+++ src/misc/enums.ts
@@ -1,20 +1,4 @@
 /**
- * Available paging position where should be rendered
- */
-export enum PagingPosition
-{
-    /**
-     * Paging will be rendered above grid content
-     */
-    Top,
-
-    /**
-     * Paging will be rendered below grid content
-     */
-    Bottom,
-}
-
-/**
  * Available plugin types
  */
 export enum GridPluginType